Add install script
We are removing the submodules and the expectation that this git
repository exists one's home directory. Instead, the script will copy
the necessary files.

We also are removing the code to install the latest the latest ruby and
nodejs versions on first start from .zshrc. We will do this in
install.sh